Create function to extract URLs
Function extractURL created. Takes in a comment from reddit and splits the comment at all the whitespaces. It then finds whether there is an image URL within the split words.

Can recognize naked links of the form 'https://imgur.com/jpg' and of the form '[https://imgur.com/jpg](https://imgur.com/jpg)' and of the form '[This is a link!](https://imgur.com/jpg)'

It recognizes jpg, png and tif file types.