Look for mention instead of keyword
This commit changes the procedure taken by the bot to identify comments that need to have their image URLs transcribed. Rather than scanning all Reddit comments for a keyword, the bot now looks at its inbox for username mentions.

This makes our bot much more scalable since it does not have to parse thousands of comments if we decide to release it to most of Reddit.