Format fix to remove markdown syntax from post/reply
The cause for the bot posting replies that still contained markdown syntax was that the bot was posting an array that stored the transcibed text. The quick fix for this was to create a function that created an empty string variable then concatenated each element in order from the array to the string variable. The function was then implemented where ever the TranscribeImages() function was called.