Have bot run continuously
Currently, the bot waits 5 seconds between inbox checks. However, this manual rate limiting is not completely necessary as PRAW automatically handles that based on the http response.

However, we may want to leave it in in order to conserve the number of requests we make so as to process requests faster when we do get them.