Add twitter integration
This commit adds twitter integration with the !Twitter flag. If the twitter flag is raised, the bot will search google for the entire tweet. If it finds a link to the tweet, it adds the link to the bottom of the post. If it cannot find the tweet it searches for any user handles in the tweet and will provide a link to those handles if found.