modified language code recognition
The bot now spilts the all the text from a request comment and places each 'word' in a list. The bot will then check to see if any of the language codes appears once or more in the list of 'words'. This is a much better process than just checking for a language code after a specific index that could lead to translating to the wrong language